+++ src/identity/signatures.rs
@@ -13,10 +13,12 @@ pub const PROOF_TYPE_ID_MINISIGN: &str = "MitraMinisignSignature2022A";
 // - Canonicalization algorithm: JCS
 // - Digest algorithm: SHA-256
 // - Signature algorithm: RSASSA-PKCS1-v1_5
-pub const PROOF_TYPE_JCS_RSA: &str = "JcsRsaSignature2022";
+pub const PROOF_TYPE_JCS_RSA: &str = "MitraJcsRsaSignature2022";
+pub const PROOF_TYPE_JCS_RSA_LEGACY: &str = "JcsRsaSignature2022";
 
 // Similar to EthereumPersonalSignature2021 but with JCS
-pub const PROOF_TYPE_JCS_EIP191: &str ="JcsEip191Signature2022";
+pub const PROOF_TYPE_JCS_EIP191: &str = "MitraJcsEip191Signature2022";
+pub const PROOF_TYPE_JCS_EIP191_LEGACY: &str ="JcsEip191Signature2022";
 
 // Similar to Ed25519Signature2020
 // https://w3c-ccg.github.io/di-eddsa-2020/#ed25519signature2020
@@ -38,8 +40,10 @@ impl FromStr for SignatureType {
     fn from_str(value: &str) -> Result<Self, Self::Err> {
         let signature_type = match value {
             PROOF_TYPE_JCS_EIP191 => Self::JcsEip191Signature,
+            PROOF_TYPE_JCS_EIP191_LEGACY => Self::JcsEip191Signature,
             PROOF_TYPE_JCS_ED25519 => Self::JcsEd25519Signature,
             PROOF_TYPE_JCS_RSA => Self::JcsRsaSignature,
+            PROOF_TYPE_JCS_RSA_LEGACY => Self::JcsRsaSignature,
             _ => return Err(ConversionError),
         };
         Ok(signature_type)